Remove TopoJSON from bundles
