--- packages/app/src/components/choropleth/logic/topology.ts
@@ -1,33 +1,0 @@
-import type { FeatureCollection, MultiPolygon, Polygon } from 'geojson';
-import * as topojson from 'topojson-client';
-import inTopology from './in.topo.json';
-import nlTopology from './nl-vr-gm.topo.json';
-
-export type CodedGeoProperties = {
-  code: string;
-};
-
-export type CodedGeoJSON = FeatureCollection<
-  MultiPolygon | Polygon,
-  CodedGeoProperties
->;
-
-export const nlGeo = topojson.feature(
-  nlTopology,
-  nlTopology.objects.nl_features
-) as CodedGeoJSON;
-
-export const vrGeo = topojson.feature(
-  nlTopology,
-  nlTopology.objects.vr_features
-) as CodedGeoJSON;
-
-export const gmGeo = topojson.feature(
-  nlTopology,
-  nlTopology.objects.gm_features
-) as CodedGeoJSON;
-
-export const inGeo = topojson.feature(
-  inTopology,
-  inTopology.objects.in_features
-) as CodedGeoJSON;
