+++ packages/app/src/components/choropleth/logic/types.ts
@@ -5,8 +5,12 @@ import type {
   VrCollection,
 } from '@corona-dashboard/common';
 import type { ParsedFeature } from '@visx/geo/lib/projections/Projection';
-import type { Feature, MultiPolygon, Polygon } from 'geojson';
-import type { CodedGeoProperties } from './topology';
+import type {
+  FeatureCollection,
+  Feature,
+  MultiPolygon,
+  Polygon,
+} from 'geojson';
 
 export type Unpack<T> = T extends infer U ? U : never;
 
@@ -106,6 +110,15 @@ export type MappedDataItem<T extends MapType> = T extends 'gm'
   : never;
 
 export type ChoroplethDataItem = GmDataItem | VrDataItem | InDataItem;
+
+export type CodedGeoProperties = {
+  code: string;
+};
+
+export type CodedGeoJSON = FeatureCollection<
+  MultiPolygon | Polygon,
+  CodedGeoProperties
+>;
 
 export type ParsedFeatureWithPath = Omit<
   ParsedFeature<Feature<MultiPolygon | Polygon, CodedGeoProperties>>,
